Add Room concept, and fix basic websockets connection
This PR contains a few things:

Built out "rooms" concept, which mostly replaces the "poll rooms".

`daphne` needed to be added to Django's INSTALLED_APPS to allow django's
manage.py runserver to serve the asgi application. This is a new feature
as of daphne 4.0, so wasn't well documented yet. See:
https://github.com/django/daphne/blob/main/CHANGELOG.txt

Removed some channels auth stuff that was failing - we aren't using
django auth at all, for now.
