+++ mathplayground/asgi.py
@@ -9,7 +9,6 @@ https://docs.djangoproject.com/en/4.1/howto/deployment/asgi/
 
 import os
 
-from channels.auth import AuthMiddlewareStack
 from channels.routing import ProtocolTypeRouter, URLRouter
 from channels.security.websocket import AllowedHostsOriginValidator
 from django.core.asgi import get_asgi_application
@@ -19,15 +18,13 @@ os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'mathplayground.settings')
 # is populated before importing code that may import ORM models.
 django_asgi_app = get_asgi_application()
 
-from mathplayground.polls import routing  # noqa
+from mathplayground.rooms import routing  # noqa
 
 application = ProtocolTypeRouter({
     'http': django_asgi_app,
     'websocket': AllowedHostsOriginValidator(
-        AuthMiddlewareStack(
-            URLRouter(
-                routing.websocket_urlpatterns
-            )
+        URLRouter(
+            routing.websocket_urlpatterns
         )
     ),
 })
